Add support for nodejs chunked requests
Many changes here, still very rough on the edges.

1. I added `doo` + `cljsbuild` to `project.clj` mostly to be able to run
nodejs tests. You will notice that I dropped the other tests... I wasn't
able to get everything working together since I need to run some tests
on the browser and other tests on nodejs. Maybe you know of a good way
of doing this?

The nice thing is that if we do manage to integrate `doo`, its going to
be a breeze to test on multiple environments and it supports "watching"
for tests, which is nice when developing.

2. I added a test case to showcase the problem. Super ugly test. It
would be much better if I could get the body that the server sent and
assert that it was what the client received. Again, not sure how to do
that, maybe you do :D

3. I added the fix to node.cljs:
Essentially I just read several times instead of just once and
when the `end` event comes in I invoke the callback.